Update GHA for better UI presentation
This commit updates the Git Hub Actions yaml config file to attempt to
separate each check as its own job. The goal of separating each check
into its own job is so that they each show clearly at the home page of
the PR instead of having to click on the GHA actions link.

Signed-off-by: Rose Judge <rjudge@vmware.com>